Added composer.lock
When you track composer.lock with your source code control system it
allows you to do a composer update on your development machine and then,
later, a composer install on your production machine. The composer
install command will make sure all packages are the correct version as
specified in the composer.lock file. Thus production uses not only the
same packages, but the same versions of the packages as your development
machine.